Redo forecasts overview page.
