+++ app/competitions/[competitionId]/forecasts/forecast-table-header.tsx
