+++ app/competitions/[competitionId]/forecasts/loading.tsx
