+++ app/competitions/[competitionId]/forecasts/user-year-selector.tsx
